requirements.txt
pycryptodome == 3.9.6
